create logs and csv

+++ Cloud/code/cloud_main.py
@@ -1,6 +1,7 @@
 import csv
 import datetime
 import logging
+import os.path
 import time
 
 
@@ -17,14 +18,18 @@ class CloudMain:
     @staticmethod
     def save_dict_to_csv(csv_file, data):
         writer = csv.writer(csv_file)
-        writer.writerow(['Picture Number', 'Number of Vehicles'])
         for key, value in data.items():
             writer.writerow([key, value])
 
     # Define a static method called "road_check" that takes in an argument called "info"
     def road_check(self, info):
         self.txt_file = open('/data/cloud_logs.txt', 'a')
-        csv_file = open('/data/count_results.csv', 'w', newline='')
+        if os.path.exists('/data/count_results.csv'):
+            csv_file = open('/data/count_results.csv', "a", newline="")
+        else:
+            csv_file = open('/data/count_results.csv', "w", newline="")
+            header = ['Picture Number', 'Number of Vehicles']
+            (csv.writer(csv_file)).writerow(header)
 
         # Check if the "info" argument is a dictionary
         if type(info) is dict:
